Add files via upload
A simple code to fitting the rotational velocity of NGC 6503 galaxy using the Navarro–Frenk–White (NFW) and Scalar Field Dark Matter (SFDM) dentity profiles. Give a plot of the rotation velocity data with the resulting fits. It also gives estimated values ​​for the radius and density of the halo core.